move shapley convenience functions to notebook support
